+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/localBackend.ts
@@ -134,7 +134,9 @@ export class LocalBackend implements Partial<backend.Backend> {
                         type:
                             projectId === LocalBackend.currentlyOpeningProjectId
                                 ? backend.ProjectState.openInProgress
-                                : backend.ProjectState.closed,
+                                : project.lastOpened != null
+                                ? backend.ProjectState.closed
+                                : backend.ProjectState.created,
                     },
                 })
             }